move special functions to utils
