+++ computer-graphics-25.vcxproj.filters
@@ -17,6 +17,9 @@
     <Filter Include="source\Scene">
       <UniqueIdentifier>{afea24a6-13c0-40d8-b345-49c47e1a6435}</UniqueIdentifier>
     </Filter>
+    <Filter Include="source\utils">
+      <UniqueIdentifier>{17e36fa8-ecaa-4a65-afaa-8075a4145b36}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <None Include=".gitignore" />
@@ -34,6 +37,9 @@
     <ClCompile Include="source\Scene\TriangleScene.cpp">
       <Filter>source\Scene</Filter>
     </ClCompile>
+    <ClCompile Include="source\utils\utils.cpp">
+      <Filter>source\utils</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="source\window\Window.h">
@@ -47,6 +53,9 @@
     </ClInclude>
     <ClInclude Include="source\Scene\TriangleScene.h">
       <Filter>source\Scene</Filter>
+    </ClInclude>
+    <ClInclude Include="source\utils\utils.h">
+      <Filter>source\utils</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
